remove some logic from WorkspaceController
